Simpler and ergonomics commit and rollback.
When you try to use commit or rollback the methods are self consuming. This is very confusing for new rust users who do not understand that the transaction variable can no longer be used after the commit or rollback method is called. Especially if you come from environments like Java or C#, where the commit or the rollback don't have this behavior.

Additionally, these self-consuming methods are more difficult to use with Arc<Mutex<DatabaseTransaction>> because it is necessary to extract the value of the Arc and the Mutex. To be able to call the Commit or the Rollback. That is much more confusing and unergonomic.

That is why alternative methods are proposed that allow commit and rollback to be used in a simpler way.