call `EnumIter::get` using fully qualified syntax (#2321)
* call `EnumIter::get` using fully qualified syntax

`itertools` version 0.13.0 added a `get` method that takes precedence
over the `EnumIter::get` method. The `EnumIter` derive should be calling
the method fully-qualified to ensure that the local version is called
rather than the method that is provided by the `Itertools` trait. This
was fixed in `strum` 0.26.x, but since this is a forked/adapted version
of the `EnumIter` derive, it requires the same fix .

* CHANGELOG

---------

Co-authored-by: Billy Chan <ccw.billy.123@gmail.com>