fix: include column_name in DeriveColumn FromStr impl (#2603)
* fix: include column_name in DeriveColumn FromStr

This updates the `FromStr` implementation in `DeriveColumn` to ensure that `column_name` attributes are recognized.

When custom `column_name`s did not match the snake_case or lowerCamelCase variants (that the derivation already provides), then a column Iden could fail to map to a column.

* test: ensure Column::from_str recognizes column_name

This adds a test case for `DeriveEntityModel` colum names to verify that `Column::from_str` succeeds on a particularly weird `column_name`.