sea-orm-rocket: update for compatibility with rocket 0.5.1
The database API for Rocket has changed between 0.5.0 and 0.5.1,
and database pools must now use rocket_db_pools. This breaks
compatibility with the current implementation of sea-orm-rocket.

Based on the working implementation at
https://github.com/amyipdev/sea-orm-rocket, this patch updates the
code to use rocket_db_pools and bumps the version number to 0.6.0
to reflect that this version is NOT compatible with the prior v0.5.5.

A maintainer with push access to crates `sea-orm-rocket` and
`sea-orm-rocket-codegen` will need to update the crates upon acceptance
of this PR.

Signed-off-by: Amy Parker <amy@amyip.net>
Fixes: #2730